refactor: Separate color scheme into dir
