+++ libs/design-system/src/color-scheme/ColorScheme.ts
@@ -1,0 +1,1 @@
+export type ColorScheme = 'light' | 'dark';
