+++ libs/design-system/src/color-scheme/index.ts
@@ -1,0 +1,3 @@
+export * from './ColorScheme';
+export * from './mapThemeStatusToColorScheme';
+export * from './useColorScheme';
