+++ libs/design-system/src/index.ts
@@ -30,6 +30,7 @@ export { NotificationBadge } from './notification-badge/NotificationBadge';
 export { Modal } from './modal/Modal';
 export { LoadingOverlay } from './loading-overlay/LoadingOverlay';
 export { NameInput } from './name-input/NameInput';
+export * from './color-scheme';
 export * from './panda';
 export * from './pagination';
 export * from './cards';
